% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baseballData}
\alias{baseballData}
\title{Dataframe of 12,145 observations of baseball games in 2010}
\format{
A data frame with 12145 rows and 5 variables:
\describe{
\item{Date}{date game was played}
\item{Home}{abbreviation for home team (i.e. stadium where game played)}
\item{Visitor}{abbreviation for visiting team}
\item{HomeScore}{Runs scored by the home team}
\item{VisitorScore}{Runs scored by the visiting team}
}
}
\usage{
baseballData
}
\description{
Dataframe of 12,145 observations of baseball games in 2010
}
\keyword{datasets}
