% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_create.R
\name{dag_create}
\alias{dag_create}
\title{Create a graph object focused on drawing a DAG.}
\usage{
dag_create()
}
\value{
a list object of class `causact_graph` consisting of 6 dataframes.  Each data frame is responsible for storing information about nodes, edges, plates, and the relationships among them.
}
\description{
Generates a \code{causact_graph} graph object that is set-up for drawing DAG graphs.
}
\examples{
# With `dag_create()` we can create an empty graph and
# add in nodes (`dag_node()`), add edges (`dag_edge`), and
# view the graph with `dag_render()`.
dag_create()
}
