% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_diagrammer.R
\name{dag_diagrammer}
\alias{dag_diagrammer}
\title{The graph object should be of class \code{causact_graph} and created using \code{dag_create()}.}
\usage{
dag_diagrammer(graph, wrapWidth = 24, shortLabel = FALSE)
}
\arguments{
\item{graph}{a graph object of class \code{causact_graph} created using \code{dag_create()}.}

\item{wrapWidth}{a required character label that describes the node.}

\item{shortLabel}{a longer more descriptive character label for the node.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
The graph object should be of class \code{causact_graph} and created using \code{dag_create()}.
}
\examples{
\dontrun{
dag_create() \%>\%
dag_node("Get Card","y",
         rhs = bernoulli(theta),
         data = carModelDF$getCard) \%>\%
  dag_diagrammer() \%>\%
  DiagrammeR::render_graph(title = "DiagrammeR Version of causact_graph")
  }
}
