% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_edge.R
\name{dag_edge}
\alias{dag_edge}
\title{Add an edge between nodes in a graph object.}
\usage{
dag_edge(graph, from, to)
}
\arguments{
\item{from}{the parent node label or description from which the edge is connected.}

\item{to}{the child node label or description from which the edge is connected.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
With a graph object of class \code{dgr_graph} created from \code{dag_create}, add an edge between nodes in the graph.
}
\examples{
# Create a graph with 2 connected nodes
dag_create() \%>\%
  dag_node("X") \%>\%
  dag_node("Y") \%>\%
  dag_edge(from = "X", to = "Y") \%>\%
  dag_render(shortLabel = TRUE)
}
