% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_node.R
\name{dag_node}
\alias{dag_node}
\title{Add a node to an existing graph object}
\usage{
dag_node(graph, descr = as.character(NA), label = as.character(NA),
  data = NULL, rhs = NA, child = as.character(NA), obs = FALSE,
  keepAsDF = FALSE, extract = as.logical(NA))
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} created using \code{dag_create()}.}

\item{label}{a required character label that describes the node.}

\item{description}{a longer more descriptive character label for the node.}

\item{distr}{a variable greta array function such as \code{uniform, normal, lognormal, bernoulli,} etc. Function arguments are optional.  Use \code{distr = NA} to suppress distribution assumptions.  Valid values include \code{greta::normal}, \code{normal}, and \code{normal(6,2)}.}

\item{observed}{a logical value indicating whether the node is observed.}

\item{formulaString}{a character string representing an operation that should be performed to calculate the node's value from its parents.  When input, \code{distr} is ignored.}

\item{children}{an optional character vector of existing node labels or node descriptions.  Directed edges from the newly created node will be created.}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
The graph object should be of class \code{dgr_graph} and created using \code{dag_create()}.
}
\examples{
# Create an empty graph and add 2 nodes by using
# the `dag_node()` function twice
graph <-
  dag_create() \%>\%
  dag_node("X") \%>\%
  dag_node("Y", children = "X")
dag_render(graph, shortLabel = TRUE)

dag_create() \%>\%
  dag_node("Y","Sales Price",
           distr = normal, observed = TRUE) \%>\%
  dag_node("X","Square Footage",
           observed = TRUE, distr = NA) \%>\%
  dag_node("mean","Exp Sales Price",
           children = "Sales Price",
           formulaString = "alpha + beta * x") \%>\%
  dag_node("beta", "Price per Sq Foot",
           children = "mean") \%>\%
  dag_node("SD", "Std Dev of Sales Price",
           distr = lognormal(0,3),
           children = "Sales Price") \%>\%
  dag_node("alpha", "Intercept",
           children = "mean")  \%>\%
  dag_edge("Square Footage","mean") \%>\%
  dag_render()
}
