% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_plate.R
\name{dag_plate}
\alias{dag_plate}
\title{Set node attribute values to reflect a desired cluster}
\usage{
dag_plate(graph, indexLabel, description, nodeLabels, data = as.character(NA))
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} created using \code{dag_create()}.}

\item{indexLabel}{a short character string to use as an index letter
nodes.}

\item{description}{a longer more descriptive label for the cluster/plate.}

\item{nodeLabels}{a vector of node labels or descriptions to include in the list of nodes}
}
\value{
a graph object of class \code{dgr_graph}.
}
\description{
From a graph object of class \code{dgr_graph}, set node attribute values for
  one or more nodes.
}
\examples{
# create small DF
dag_create() \%>\%
    dag_node("Y","House Prices") \%>\%
    dag_plate("i","Observation",nodeLabels = "Y") \%>\%
    dag_render(shortLabel = TRUE)
}
