% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagp_plot.R
\name{dagp_plot}
\alias{dagp_plot}
\title{Plot posterior distribution of latent parameters in \code{causact_graph} model.}
\usage{
dagp_plot(drawsDataFrame)
}
\arguments{
\item{drawsDataFrame}{one of the two posterior data frame objects created by \code{dag_greta()}.  If \code{drawsDataFrame=drawsDF}, each parameter gets its own density plot.  If \code{drawsDataFrame=tidyDrawsDF} (recommended usage), parameters are grouped into facets based on whether they share the same prior or not.  10 and 90 percent credible intervals are displayed for the posterior distributions.}
}
\value{
a credible interval plot of all latent posterior distribution parameters.
}
\description{
The graph object should be of class \code{causact_graph} and created using \code{dag_create()}.
}
\examples{
library(greta)
# Create a 2 node graph
graph = dag_create() \%>\%
  dag_node("Get Card","y",
         rhs = bernoulli(theta),
         data = carModelDF$getCard) \%>\%
  dag_node(descr = "Card Probability by Car",label = "theta",
           rhs = beta(2,2),
           child = "y")
graph \%>\% dag_render()
graph \%>\% dag_greta(mcmc=TRUE)
drawsDF \%>\% dagp_plot()
tidyDrawsDF \%>\% dagp_plot()

# multiple plate example
library(dplyr)
poolTimeGymDF = gymDF \%>\%
mutate(stretchType = ifelse(yogaStretch == 1,
                            "Yoga Stretch",
                            "Traditional")) \%>\%
group_by(gymID,stretchType,yogaStretch) \%>\%
  summarize(nTrialCustomers = sum(nTrialCustomers),
            nSigned = sum(nSigned))
graph = dag_create() \%>\%
  dag_node("Cust Signed","k",
           rhs = binomial(n,p),
           data = poolTimeGymDF$nSigned) \%>\%
  dag_node("Probability of Signing","p",
           rhs = beta(2,2),
           child = "k") \%>\%
  dag_node("Trial Size","n",
           data = poolTimeGymDF$nTrialCustomers,
           child = "k") \%>\%
  dag_plate("Yoga Stretch","x",
            nodeLabels = c("p"),
            data = poolTimeGymDF$stretchType,
            addDataNode = TRUE) \%>\%
  dag_plate("Observation","i",
            nodeLabels = c("x","k","n")) \%>\%
  dag_plate("Gym","j",
            nodeLabels = "p",
            data = poolTimeGymDF$gymID,
            addDataNode = TRUE)
graph \%>\% dag_render()
graph \%>\% dag_greta(mcmc=TRUE)
tidyDrawsDF \%>\% dagp_plot()
}
