% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{delivDF}
\alias{delivDF}
\title{117,790 line items associated with 23,339 shipments.}
\format{
A data frame (tibble) with 117,790 rows and 5 variables:
\describe{
\item{shipID}{unique ID for each shipment}
\item{plannedShipDate}{shipment date promised to customer}
\item{actualShipDate}{date the shipment was actually shipped}
\item{partID}{unique part identifier}
\item{quantity}{quantity of partID in shipment}
}
}
\source{
Adam Fleischhacker
}
\usage{
delivDF
}
\description{
A dataset containing the line items, mostly parts, asssociated
with 23,339 shipments from a US-based warehouse.
}
\keyword{datasets}
