% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dot2.R
\name{generate_dot2}
\alias{generate_dot2}
\title{Generate DOT code using a graph object}
\usage{
generate_dot2(graph)
}
\value{
a character vector of length 1 containing Graphviz DOT code.
}
\description{
Generates Graphviz DOT code as an R character object using DiagrammeR graph
  object.
}
