% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gymDF}
\alias{gymDF}
\title{Dataframe of 44 observations of free crossfit classes data
Each observation indicates how many students that participated in the free
month of crossfit signed up for the monthly membership afterwards}
\format{
A data frame with 44 rows and 5 variables:
\describe{
\item{gymID}{unique gym identifier}
\item{nTrialCustomers}{number of unique customers taking free trial classes}
\item{nSigned}{number of customers from trial that sign up for membership}
\item{yogaStretch}{whether trial classes included a yoga type stretch}
\item{timePeriod}{month number, since inception of company, for which trial period was offered}
}
}
\usage{
gymDF
}
\description{
Dataframe of 44 observations of free crossfit classes data
Each observation indicates how many students that participated in the free
month of crossfit signed up for the monthly membership afterwards
}
\keyword{datasets}
