% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{houseDFDescr}
\alias{houseDFDescr}
\title{Dataframe of 523 descriptions of data values from "The Ames Housing dataset", compiled by Dean De Cock for use in data science education.
Each observation is a possible value from a variable in the \code{houseDF} dataset.}
\format{
A data frame with 260 rows and 2 variables:
\describe{
\item{varName}{the name and description of a variable stored in the \code{houseDF} dataset}
\item{varValueDescr}{The value and accompanying interpretation for values in the \code{houseDF} dataset}
}
}
\source{
Accessed Jan 22, 2019. Kaggle dataset on "House Prices: Advanced Regression Techniques".
}
\usage{
houseDFDescr
}
\description{
Dataframe of 523 descriptions of data values from "The Ames Housing dataset", compiled by Dean De Cock for use in data science education.
Each observation is a possible value from a variable in the \code{houseDF} dataset.
}
\keyword{datasets}
