% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_causact_deps.R
\name{install_causact_deps}
\alias{install_causact_deps}
\title{Install causact's python dependencies like numpyro, arviz, and xarray.}
\usage{
install_causact_deps()
}
\description{
\code{install_causact_deps()} installs python, the numpyro and arviz packages, and their
direct dependencies.
}
\details{
You may be prompted to download and install miniconda if reticulate
did not find a non-system installation of python. Miniconda is the
only supported installation method for users, as it ensures that the R
python installation is isolated from other python installations. All python
packages will by default be installed into a self-contained conda or venv
environment named "r-causact". Note that "conda" is the only supported method for install.

If you initially declined the miniconda installation prompt, you can later
manually install miniconda by running \code{\link[reticulate:install_miniconda]{reticulate::install_miniconda()}}.

If you manually configure a python environment with the required
dependencies, you can tell R to use it by pointing reticulate at it,
commonly by setting an environment variable:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{Sys.setenv("RETICULATE_PYTHON" = "~/path/to/python-env/bin/python")
}\if{html}{\out{</div>}}
}
