% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{The magrittr pipe}
\value{
Pipe a value forward into a function- or call expression and return the function on the \code{rhs} with the \code{lhs} used as the first argument.
}
\description{
causact uses the pipe function,
\verb{\\\%>\\\%} to turn function composition into a
series of imperative statements.
}
