% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_in_spec2.R
\name{replace_in_spec2}
\alias{replace_in_spec2}
\title{Razor-like template for diagram specification}
\usage{
replace_in_spec2(spec)
}
\arguments{
\item{spec}{string spec to be parsed and evaluated.}
}
\description{
Use Razor-like syntax to define a template for use in a \code{grViz} diagram.
}
