% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schoolsDF}
\alias{schoolsDF}
\title{This example, often referred to as 8-schools, was popularized by its inclusion in Bayesian Data Analysis (Gelman, Carlin, & Rubin 1997).}
\format{
A data frame with 8 rows and 3 variables:
\describe{
  \item{y}{estimated treatment effect at a particular school}
  \item{sigma}{standard error of the treamtment effect estimate}
  \item{schoolName}{an identifier for the school represented by this row}
  }
}
\usage{
schoolsDF
}
\description{
This example, often referred to as 8-schools, was popularized by its inclusion in Bayesian Data Analysis (Gelman, Carlin, & Rubin 1997).
}
\keyword{datasets}
