% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ticketsDF}
\alias{ticketsDF}
\title{Dataframe of 55,167 observations of the number of tickets
written by NYC precincts each day
Data modified from https://github.com/stan-dev/stancon_talks/tree/master/2018/Contributed-Talks/01_auerbach which originally sourced data from https://opendata.cityofnewyork.us/}
\format{
A data frame with 55167 rows and 4 variables:
\describe{
  \item{precinct}{unique precinct identifier representing precinct of issuing officer}
  \item{date}{the date on which ticket violations occurred}
  \item{month_year}{the month_year extracted from date column}
  \item{daily_tickets}{Number of tickets issued out of precinct on this day}
}
}
\usage{
ticketsDF
}
\description{
Dataframe of 55,167 observations of the number of tickets
written by NYC precincts each day
Data modified from https://github.com/stan-dev/stancon_talks/tree/master/2018/Contributed-Talks/01_auerbach which originally sourced data from https://opendata.cityofnewyork.us/
}
\keyword{datasets}
