% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causact-package.R
\docType{package}
\name{causact-package}
\alias{causact}
\alias{causact-package}
\title{causact: Fast, Easy, and Visual Bayesian Inference}
\description{
Accelerate Bayesian analytics workflows in 'R' through interactive modelling, visualization, and inference. Define probabilistic graphical models using directed acyclic graphs (DAGs) as a unifying language for business stakeholders, statisticians, and programmers. This package relies on interfacing with the 'numpyro' python package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/flyaflya/causact}
  \item \url{https://www.causact.com/}
  \item Report bugs at \url{https://github.com/flyaflya/causact/issues}
}

}
\author{
\strong{Maintainer}: Adam Fleischhacker \email{ajf@udel.edu} [copyright holder]

Other contributors:
\itemize{
  \item Daniela Dapena [contributor]
  \item Rose Nguyen [contributor]
  \item Jared Sharpe [contributor]
}

}
\keyword{internal}
