% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corruptDF}
\alias{corruptDF}
\title{Dataframe of 174 observations where information on the human developmet index (HDI)
and the corruption perceptions index (CPI) both exist.
Each observation is a country.}
\format{
A data frame with 174 rows and 7 variables:
\describe{
\item{country}{country name}
\item{region}{region name as given with CPI rating}
\item{countryCode}{three letter abbreviation for country}
\item{regionCode}{four letter or less abbreviation for country}
\item{population}{2017 country population}
\item{CPI2017}{The Corruption Perceptions Index score for 2017: A country/territory’s score indicates the perceived level of public sector corruption on a scale of 0-100, where 0 means that a country is perceived as highly corrupt and a 100 means that a country is perceived as very clean. }
\item{HDI2017}{The human development index score for 2017: the Human Development Index (HDI) is a measure of achievement in the basic dimensions of human development across countries.  It is an index made from a simple unweighted average of a nation’s longevity, education and income and is widely accepted in development discourse.}
}
}
\source{
CPI data available from https://www.transparency.org/en/cpi/2017. Accessed Feb 24, 2024. Consumer Perception Index 2017 by Transparency International is licensed under CC-BY- ND 4.0.

\url{https://hdr.undp.org/data-center/human-development-index#/indicies/HDI} HDA data accessed on Oct 1, 2018.

\url{https://data.worldbank.org/} Population data accessed on Oct 1, 2018.
}
\usage{
corruptDF
}
\description{
Dataframe of 174 observations where information on the human developmet index (HDI)
and the corruption perceptions index (CPI) both exist.
Each observation is a country.
}
\keyword{datasets}
