% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_create.R
\name{dag_create}
\alias{dag_create}
\title{Create a graph object focused on drawing a DAG.}
\usage{
dag_create(...)
}
\arguments{
\item{...}{Any arguments to pass to \code{DiagrammeR::create_graph()}.}
}
\description{
Generates a \code{dgr_graph} graph object that is set-up for drawing DAG graphs.
}
\examples{
# With `dag_create()` we can
# simply create an empty graph (and
# add in nodes and edges later
# with other functions)
dag_create() \%>\% dag_render(shortLabel = TRUE)
}
