% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_greta.R
\name{dag_greta}
\alias{dag_greta}
\title{Create a graph object focused on drawing a DAG.}
\usage{
dag_greta(graph)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph} created using \code{dag_create()} with data argument populated.}
}
\description{
The graph object should be of class \code{dgr_graph} and created using \code{dag_create()}.
}
\examples{
# With `dag_create()` we can
# simply create an empty graph (and
# add in nodes and edges later
# with other functions)
dag_create() \%>\% dag_render(shortLabel = TRUE)
}
