% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_render.R
\name{dag_render}
\alias{dag_render}
\title{Render the graph in various formats}
\usage{
dag_render(graph, shortLabel = FALSE, wrapWidth = 24, ...)
}
\arguments{
\item{graph}{a graph object of class \code{dgr_graph}.}

\item{shortLabel}{a logical value.  If set to \code{TRUE}, distribution and formula information is suppressed.}

\item{...}{arguments that can be passed onto \code{Diagrammer::create_graph()}.}
}
\description{
Using a \code{dgr_graph} object, render the graph in the RStudio Viewer.
}
\examples{
# Render an  empty graph
dag_create() \%>\%
  dag_render()
}
