% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulliDist.R
\name{rbern}
\alias{rbern}
\title{Make dbern,pbern,qbern,rbern available for
implementation of the Bernoulli distribution functions.}
\usage{
rbern(n, prob)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}

\item{prob}{probability of success of each trial}
}
\value{
A vector of 0's and 1's representing failure and success.
}
\description{
Density, distribution function, quantile function and random generation for the benoulli distribution with parameter \code{prob}.
}
\examples{
#Return a random result of a Bernoulli trial given \code{prob}.
rbern(n =1, prob = 0.5)
}
