% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setWorkflowDefaults.R
\name{setDirectedGraphTheme}
\alias{setDirectedGraphTheme}
\title{Set DiagrammeR defaults for graphical models}
\usage{
setDirectedGraphTheme(dgrGraph)
}
\arguments{
\item{dgrGraph}{A DiagrammeR graph}
}
\value{
An updated version of \code{dgrGraph} with good defaults for
  graphical models.
}
\description{
\code{setDirectedGraph} returns a graph with good defaults.
}
\examples{
graphA \%>\% setDirectedGraphTheme() \%>\% DiagrammeR::render_graph()
}
