% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setWorkflowDefaults.R
\name{setDirectedGraphTheme}
\alias{setDirectedGraphTheme}
\title{Set DiagrammeR defaults for graphical models}
\usage{
setDirectedGraphTheme(dgrGraph)
}
\arguments{
\item{dgrGraph}{A DiagrammeR graph}
}
\value{
An updated version of \code{dgrGraph} with good defaults for
  graphical models.
}
\description{
\code{setDirectedGraph} returns a graph with good defaults.
}
\examples{
\dontrun{
library(DiagrammeR)
create_graph() \%>\% add_node() \%>\% render_graph()  # default DiagrammeR aesthetics
create_graph() \%>\% add_node() \%>\% setDirectedGraphTheme() \%>\% render_graph() ## causact aesthetics
}
}
